requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==1.10.12
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
bcrypt==4.0.1
requests==2.31.0
python-multipart

